allow nc-server to be built inside nidas source build
Detect when the nc-server source directory is being built inside a nidas
source build, and use the nidas Environment instead.  Accordingly, make
variant arch builds work, use the nidas tool to get the nidas_util library
whether in the source tree or not, and fix the nc_server_client tool so it
works for the nidas build.