mods to work on casper with modules
Compilers and libraries are installed with module, and that depends on
environment settings, so it's easiest to pass all the environment
settings through to the SCons environment.  Also, have to be careful
not to replace PKG_CONFIG_PATH, so extend it instead.  netcdf include
path and libraries depend on module loaded, so use pkg-config to find
those settings until netcdf tool supports it.

load prefixoptions so OPT_PREFIX can be set to find netcdf C++
library, since that is not available as a module.